0.5a
Licenses are linked to & included, and credit is given where credit is
due.

Pim is now distributed under the BSD license (see README for more
information).

Added functionality to the LightingSystem:
One are now able to call the method:
setSmoothShadows(true); on any layer with an instantiated
LightingSystem, and it will smoothen the lights.
The smoothening (Gaussian Blur) is creating a few artifacts, and will be
reworked to completion at a future date.