Added Doxy doc
Please follow the format used in PimGameControl.h for doxygen
documentation.

The class description should be placed directly after the namespace
declaration, and the method descriptions should be added after the
class declaration, but before the namespace is terminated.

Not all methods need be documented, but if the method in any way is
ambigous or not descriptive enough in itself, said method should be
documented to provide clarity.